io_uring: give the iovec reference collector the same scope as the IovArray
Motivation:

IovArrayReferenceCollector was a nested class the stream unsafe held one of per
channel, while the IovArray it fills is owned once per IoUringIoHandler. The
collector grows its backing array to the high water mark of the largest writev it
ever saw and never shrinks it, so a channel that once wrote 512 iovecs keeps a
ReferenceCounted[512] for as long as it lives. Clearing the dropped entries on
reset() bounds what those slots point at, but only until the next writev: a channel
that writes once and then goes idle holds its last write's buffers indefinitely.

Modification:

Promote the collector to a top-level class, like every other reusable helper in this
package, and let IoUringIoHandler own one alongside its IovArray. The IovArray comes
in through the constructor, so reset() takes no argument. The channel reaches the
collector through the handler it already has in hand.

This is safe because nothing holds the collector across a submit. Between reset()
and the record(...) that copies its array into a write-operation slot, no other
channel's code runs: the only call in that window that reaches io_uring is
submitAndGetNow(), which enters the kernel without dispatching completions. Slots
copy rather than adopt, so a later writev overwriting the collector cannot disturb a
write that is still in flight. The class javadoc states that window, along with why a
count() delta is the right test for whether a buffer needs retaining.

Result:

One collector per event loop instead of one per channel, and an idle channel holds no
write references at all.
