+++ transport-classes-io_uring/src/main/java/io/netty/channel/uring/IoUringIoHandler.java
@@ -72,6 +72,7 @@ public final class IoUringIoHandler implements IoHandler {
     private final ByteBuffer timeoutMemory;
     private final long timeoutMemoryAddress;
     private final IovArray iovArray;
+    private final IovArrayReferenceCollector iovArrayReferenceCollector;
     private final MsgHdrMemoryArray msgHdrMemoryArray;
     private long eventfdReadSubmitted;
     private boolean eventFdClosing;
@@ -145,6 +146,7 @@ public final class IoUringIoHandler implements IoHandler {
         timeoutMemory = timeoutMemoryCleanable.buffer();
         timeoutMemoryAddress = Buffer.memoryAddress(timeoutMemory);
         iovArray = new IovArray(IoUring.NUM_ELEMENTS_IOVEC);
+        iovArrayReferenceCollector = new IovArrayReferenceCollector(iovArray);
         msgHdrMemoryArray = new MsgHdrMemoryArray((short) 1024);
     }
 
@@ -768,6 +770,14 @@ public final class IoUringIoHandler implements IoHandler {
             assert iovArray.count() == 0;
         }
         return iovArray;
+    }
+
+    /**
+     * Returns the {@link IovArrayReferenceCollector} paired with {@link #iovArray()}. A plain getter: callers are
+     * expected to have already called {@link #iovArray()} to make room, so this must not itself submit-and-clear.
+     */
+    IovArrayReferenceCollector iovArrayReferenceCollector() {
+        return iovArrayReferenceCollector;
     }
 
     MsgHdrMemoryArray msgHdrMemoryArray() {
